+++ src/exxosversion.h
@@ -5,5 +5,5 @@
  */
 #ifndef EXXOSVERSION_H
 #define EXXOSVERSION_H
-#define EXXOS_EDITION_VERSION "1.9.22"
+#define EXXOS_EDITION_VERSION "1.9.23"
 #endif